Strings-Count Palindromic Substrings
Take as input S, a string. Write a program that gives the count of substrings of this string which are palindromes and Print the ans.

Input Format
Single line input containing a string

Constraints
Length of string is between 1 to 1000.

Output Format
Integer output showing the number of palindromic substrings.

Sample Input
abc
Sample Output
3
Explanation
For the given sample case , the palindromic substrings of the string abc are "a","b" and "c".So, the ans is 3.