Count , Remove and Replace hi
Take as input str, a string. a.Write a recursive function which counts the number of times ‘hi’ appears in the string. Print the value returned. b.Write a recursive function which removes all ‘hi’ in the string. Print the value returned. c.Write a recursive function which replaces all ‘hi’ in the string with ‘bye’. Print the value returned.

Input Format
Enter a string

Constraints
None

Output Format
Display the no. of 'hi', string without 'hi' , string in which 'hi' is replaced with 'bye'

Sample Input
abchibi
Sample Output
1
abcbi
abcbyebi